mybatis part 2 implementing DAO classes

+++ src/main/java/db/dao/mysql/OrderDAO.java
@@ -80,7 +80,7 @@ public class OrderDAO implements IOrderDAO {
         return order;
     }
 
-    public Order getOrderByPayment(int dbPayment) {
+    public Order getOrderByPayment(double dbPayment) {
         Order order = null;
         Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM orders WHERE payment = (?)";
